+++ core/middleware/incoming_payloads/runtime/motion_command_payload.hpp
@@ -9,8 +9,8 @@ namespace middleware_incoming_payloads
   struct motion_command_payload_data
   {
     bool drive_enabled = false;
-    std::int16_t linear_velocity_mm_s = 0;
-    std::int16_t yaw_rate_mdeg_s = 0;
+    std::int32_t linear_velocity_mm_s = 0;
+    std::int32_t yaw_rate_mdeg_s = 0;
     std::uint32_t received_time_ms = 0U;
   };
